refactor: relocated functions
